extract conditions matcher in a module
